+++ src/backend/headless_cli.rs
@@ -1,17 +1,14 @@
-use crate::backend::desm_mapping::apply_desm_mapping;
-use crate::backend::preload_p2::get_json;
-use crate::backend::repository::Repository;
-use crate::backend::transformations::Transformation;
+use crate::backend::init_conversion::load_mapping_file;
 use crate::p2_p3_common::create_output_files;
 use crate::state::{AppState, Mapping};
 use crate::trace_dbg;
 
 use clap::Parser;
-use serde_json::json;
-use std::collections::HashMap;
 use std::fs::read_dir;
 use std::io::Result;
 use std::path::Path;
+
+use super::init_conversion::load_input_file;
 
 pub fn run_headless(cli_args: &mut Args, state: &mut AppState) -> Result<()> {
     check_args(cli_args)?;
@@ -92,28 +89,8 @@ pub fn check_args(cli_args: &Args) -> Result<()> {
 }
 
 pub fn load_files_apply_transformations(state: &mut AppState) {
-    state.repository = Repository::from(HashMap::from_iter(vec![
-        (
-            state.mapping.input_format(),
-            get_json(&state.input_path).expect("No source file found"),
-        ),
-        (state.mapping.output_format(), json!({})),
-    ]));
-
-    trace_dbg!("Successfully loaded the input file");
-
-    if state.mapping_path == "DESM" {
-        apply_desm_mapping(state);
-    } else {
-        // todo: remove unwrap
-        let rdr = std::fs::File::open(&state.mapping_path).unwrap();
-        let transformations: Vec<Transformation> = serde_json::from_reader(rdr).unwrap();
-
-        trace_dbg!("Successfully loaded the mapping file");
-
-        state.repository.apply_transformations(transformations, state.mapping);
-    }
-
+    load_input_file(state, true);
+    load_mapping_file(state);
     create_output_files(state);
 }
