+++ src/backend/mod.rs
@@ -1,11 +1,11 @@
 pub mod candidate_value;
 pub mod desm_mapping;
+pub mod getters_resolvers;
 pub mod headless_cli;
+pub mod init_conversion;
 pub mod jsonpointer;
 pub mod leaf_nodes;
 pub mod logging;
-pub mod preload_p2;
 pub mod repository;
-pub mod resolve;
 pub mod transformations;
 pub mod update_display;
