+++ src/events/p2_p3_common.rs
@@ -5,7 +5,7 @@ use std::path::Path;
 use std::{char, fs::File};
 
 use super::{is_mouse_over_area, p2_handler::clear_progress};
-use crate::backend::preload_p2::get_json;
+use crate::backend::init_conversion::get_json;
 use crate::backend::update_display::{update_display_section, update_pointer};
 use crate::{
     backend::{
