+++ src/render/p2.rs
@@ -1,5 +1,5 @@
 use crate::{
-    backend::resolve::value_to_str,
+    backend::getters_resolvers::value_to_str,
     mapping_bars::{render_manytoone_bar, render_mapping_bar},
     popups::{render_popup_exit_warning, render_popup_mapping, render_popup_uncompleted_warning_p2},
     state::{translate, AppState, MappingOptions, P2P3Tabs},
